+++ packages/vue-app-layout/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @fastkit/vue-app-layout
 
+## 0.14.8
+
+### Patch Changes
+
+- Various Typo corrections were made.
+
+- Updated dependencies []:
+  - @fastkit/vue-scroller@0.14.8
+
 ## 0.14.7
 
 ### Patch Changes